move QgsMultiPolygon tests from QgsGeometry
